fix sqrt

+++ WSim/Source/WSimCommon/LevelSetHelper.cpp
@@ -4,6 +4,7 @@
 #include <queue>
 #include <unordered_set>
 #include <vector>
+#include <cmath>
 
 void LevelSetHelper::initializeToSphere(float *levelSet, Vec3 levelSetSize, float sphereRadius) {
     const FloatVec3 origin{
@@ -21,7 +22,7 @@ void LevelSetHelper::initializeToSphere(float *levelSet, Vec3 levelSetSize, Floa
                 const float distanceFromOriginX = x - sphereOrigin.x;
                 const float distanceFromOriginY = y - sphereOrigin.y;
                 const float distanceFromOriginZ = z - sphereOrigin.z;
-                const float distanceFromOrigin = std::sqrtf(distanceFromOriginX * distanceFromOriginX + distanceFromOriginY * distanceFromOriginY + distanceFromOriginZ * distanceFromOriginZ);
+                const float distanceFromOrigin = std::sqrt(distanceFromOriginX * distanceFromOriginX + distanceFromOriginY * distanceFromOriginY + distanceFromOriginZ * distanceFromOriginZ);
                 const float levelSetValue = distanceFromOrigin - sphereRadius;
 
                 const auto indexInLevelSet = IndexHelper::idx(x, y, z, levelSetSize);
